+++ Runtime/Scripts/Timeline/Internal/ExportPlan.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8544b2de67d0c8046b1487a23a09795d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
