+++ Runtime/Scripts/Timeline/Internal/Track.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1302926dee570b4fab70ab4421cd675
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
